[#15]refactor: model NoArgsConstructor로 통일
