+++ src/main/java/soma/edupilms/classroom/models/MyClassroomRoleResponse.java
@@ -3,7 +3,7 @@ package soma.edupilms.classroom.models;
 import java.util.List;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import soma.edupilms.guest.models.response.ClassroomAccountRole;
+import soma.edupilms.guest.models.ClassroomAccountRole;
 
 @Getter
 @NoArgsConstructor
