+++ src/main/java/soma/edupilms/guest/models/ClassroomAccountResponse.java
@@ -1,0 +1,14 @@
+package soma.edupilms.guest.models;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class ClassroomAccountResponse {
+
+    private Long accountId;
+    private Long classroomId;
+    private ClassroomAccountRole role;
+
+}
