+++ src/main/java/soma/edupilms/guest/models/ClassroomAccountRole.java
@@ -1,4 +1,4 @@
-package soma.edupilms.guest.models.response;
+package soma.edupilms.guest.models;
 
 import lombok.Getter;
